Bank Term Deposit Marketing
Bank marketing is known for its nature of developing a unique brand image, which is
treated as the capital reputation of the financial academy. It is very important for a bank to
develop good relationship with valued customers accompanied by innovative ideas which can
be used as measures to meet their requirements.