Supply Chain - OrderFraudOrNotPrediction
Supply chain fraud is becoming more common with the digitization of
business operations and ecommerce transactions. Fraud in the supply chain can
take many forms, which pose a serious threat to businesses. In some cases, it can
result in the loss of merchandise or the theft of confidential information. In other
cases, it can lead to the disruption of supply chains and the waste of resources.
With the network of intermediaries across the globe, organizations are more
vulnerable to supply chain fraud in countries with lax or less stringent rules
governing misconduct. The increase in cross-border transactions presents a
higher risk of fraudsters using phony invoices or other false documents to commit
fraud. As top management shifts its focus to handling urgent operational
concerns, prospective fraudsters see this as an opportunity to infiltrate the supply
chain.